implement new collision order
We now generate the nucleus configuration, color charge density and
Wilson lines with respect to origin for both the projectile and target
nucleus. Then we sample impact parameter vector and compute the
collision geometry assume A shifts to +vec{b}/2 and B shifts to
-vec{b}/2 in the transverse plane.

If there is no collision, we will resample the impact parameter vec{b}
and compute the geometry quantities again. We don't need to resample the
nucleon and sub-nucleonic configurations as well as generating Wilson
lines multiple times.

This setup can ensures in the input Wilson lines are always defined with
respect to the origin of the lattice. We can have a unified treatment
for generating Wilson lines on the fly, read Wilson lines from external
file, and feed Wilson lines from JIMWLK evolution.